move the testing code to dir under torch_dipu
