+++ dipu/tests/python/individual_scripts/test_dipu_op_register.py
@@ -1,6 +1,6 @@
 # Copyright (c) 2023, DeepLink.
 from multiprocessing import Process
-from tests.python.utils.local_eviron import local_eviron
+from torch_dipu.testing._internal.local_eviron import local_eviron
 
 
 def _test_op_register(mode):
