+++ dipu/torch_dipu/testing/_internal/stdout_redirector.py
